Everything ready for benchmarking
Removed old result files from previous iterations of code that
are no longer relevant. We'll be producing new results now.

Restructed the scripts directory to organize the various scripts.

Modified the runTests scripts:
We now have two versions, one for the dynamic sorting code (i.e.,
spawn on demand, the original code) and one for the network sorting
code (i.e., the new code). An extra argument has been added to these
scripts: --exe. Use this to supply the path to the executable to be
ran. I also changed it so the data set directory is assumed to be
a single directory with all of the input files in it (so no more
small/ medium/ large/). To run the scripts for the simulator:

./runTests_EMU_dynamic.sh --exe=/path/to/bitonic_dynamic.mwx --mode=SIM --datasetDir=/path/to/data

./runTests_EMU_network.sh --exe=/path/to/bitonic_network.mwx --mode=SIM --datasetDir=/path/to/data

The difference between the dynamic and network versions of the script is that
the network will sweep over combinations of nodelets and threadlets.

Made emu.mk the normal Makefile. This is how we can use it for our purposes:

*** As of now, the only time starttiming() is NOT used is when DEBUG-=1

To compile executables with debugging:

This enables input and output array printing, as well as checking whether the output is correctly sorted.

make all DEBUG=1

The starttiming() feature is not used here, so no cycle counts or detailed statistics can be obtained. The purpose of running in debug mode is, of course, debugging. Use this to verify that any changes made to the code have not broken the algorithm.

To compile executables for use on the simulator or hardware to gather cycle counts:

This enables starttiming() and prints out total cycles at the end. It does not do any printing of the in/out arrays.

make all

This will give the simulator output more details. For the hardware, it will simply give us the cycle counts. THIS IS WHAT YOU SHOULD RUN WHEN USING THE BENCHMARK SCRIPTS TO GATHE RESULTS FOR BOTH THE HARDWARE AND SIMULATOR.

To compile executables for the simulator to get detailed stats but no cycle counts:

This allows for very fast and easy execution on the simulator, gather detailed stats via starttiming(), but it exits immediately after the sorting is done. So we don’t display any timing info. For our particular purposes, this can be considered more or less depreciated.

make all SIM=1