Added sim results so far (as of 5/12)
The network sorting tests failed for some runs on 2 and 4
elements when using more than 2 or 4 nodes, respectively.

I believe this is quite clear: if we have 2 elements, it
doesn't make sense to use more than 2 nodes. Likewise, if
we are sorting 4 elements, we shouldn't use more than 4 nodes.
This is fine, as we'll never show the results of sorting such
a small number of elements.